requirements.txt
Flask==0.11.1
requests==2.12.4
easydict==1.7
PyYAML==3.12
pandas==0.21.0
